+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Fix server crash when effect has rotateFace: 16
 - Fix server crash on events with ig.Action data
 - Fix ig.ENTITY.ConditionalLight
+- Unregister dummy.PlayerModel from ig.Vars#varAccessors on entity kill
 
 ## [0.5.5] 2025-10-29